Increase logging in OffscreenCanvas preserveDrawingBuffer test.
Display the canvases that are being tested in the web page, in case
their eventual display is affecting the browser's composition paths.

Increase the timeout for the preserveDrawingBuffer:false test to 2000
ms in order to see whether that eliminates flakiness seen on automated
testing machines.

Related to http://crbug.com/979843 .