Accept all NaNs when sampling invalid LDR ASTC non-sRGB
When sampling invalid LDR ASTC textures which are not sRGB, hardware is
allowed to return either the normal error color (magenta) or all NaNs.
This modifies the ASTC test to accept an all NaN result in this case.

This fixes the test on some Qualcomm Android devices including Pixel
phones.