Update test utils to use video.rVFC
The startPlayingAndWaitForVideo function uses various heuristics to
determine whether or not a video has started playing. These heuristics
mostly rely on currentTime progress, but don't aren't necessarily
reliable when determining wether there is actually new video data to
be consumed.

This commit adds cross browser support for the
video.requestVideoFrameCallback API, which can be used to replace the
heuristics, and potentially deflake tests.