Test both alpha:true and alpha:false in webgl_canvas tests. (#3071)
Remove explicit use of Promises in the inner loop, and change to use
async/await, to provide better control of the execution order of the
test cases.

Verified in Chrome, Firefox and Safari on macOS, and Safari on iOS
Simulator. ToT of all browsers pass the new tests; Firefox has a
couple of preexisting failures with the
RGBA10_A2/RGBA/UNSIGNED_INT_2_10_10_10_REV combination.