Update test utils to use video.rVFC (#3065)
* Update test utils to use video.rVFC

The startPlayingAndWaitForVideo function uses various heuristics to
determine whether or not a video has started playing. These heuristics
mostly rely on currentTime progress, but aren't necessarily
reliable when determining whether there is actually new video data to
be consumed.

This PR adds cross browser support for the
video.requestVideoFrameCallback API, which can be used to replace the
heuristics, and potentially deflake tests. It applies to the 1.0.1, 1.0.2, 1.0.3,
and 2.0.0 conformance test suite snapshots, as well as top-of-tree.