Choose which cross-origin URL to use up-front.
The earliest version of this code, which was robust, made the
determination of whether to use a local or remote URL cor cross-origin
images up front.  Network timeouts from machines that deliberately
aren't connected to the open Internet can't be relied upon as a
fallback mechanism. Change the code back to examine the test's URL and
choose whether to load local or remote images based on that.

Follow-on to #3052 and #3077.

Should fix #3076 conclusively.