Test JPEG with linear gamma in image_bitmap_from_blob tests.
Restructure the tests to use the Fetch API and Promises to make this
addition easier to read.

Make tolerance configurable, and use a large value for compatibility
with Firefox Nightly on macOS.

The new JPEG resource was exported from the included Photoshop
document which was constructed from scratch for this test case, based
on the original report from Figma.

Regression test for http://crbug.com/1058160 .