Mark WebGL 2.0 Compute specification as obsolete. (#3261)
Automatic regeneration of the index.html from the Bikeshed sources is
giving developers the impression that this specification is being
actively developed. Add a prominent note in the abstract that the
WebGPU spec is the path forward for compute shaders on the web, and
that this one is obsolete, while thanking Intel for their tremendous
contributions.