Don't use await for video.play() promise.
Some tests are finishing as soon as the first frame has come in
and triggering a pause() or other condition which thorws a
'AssertionError: Video failed to play(): AbortError: The play()
request was interrupted by a call to pause().' error.

So just check for no sources and use the error handler like we
do on older WebGL test versions.