Add WebCodecs VideoFrame as a TexImageSource.
Adding VideoFrame as a TexImageSource allows texImage*() methods to
interoperate with VideoFrames. VideoFrames work with texImage*() the
same way the existing point-in-time capture for HTMLVideoElement does.

This adds a non-normative reference on the WebCodecs spec for the
VideoFrame interface and associated properties to accomplish this.

Note: VideoFrame objects are always same-origin, so I did not add
any text around origin restrictions for video frames.

https://www.w3.org/TR/webcodecs/#videoframe-interface

Bug: w3c/webcodecs#158