Fix several compatibility issues
platform.dist() is deprecated for python 3.9, so lsb_release is used
instead to get Linux version.
Browsers on Windows should be 64-bit, so PROGRAMFILES, instead of
PROGRAMFILES(X86), is used for their paths.
options, instead of chrome_options, is used as parameter of
selenium.webdriver.Chrome() to suppress the deprecation warning.