Test multisampled framebuffer draws / copies / blends.
Mimics the call structure in the following two Three.js demos:
  https://threejs.org/examples/webgl_framebuffer_texture
  https://threejs.org/examples/webgl_decals

Draws a solid color quad to the center of the viewport, copies a
portion of it to a texture, and blends that texture with the contents
of the framebuffer.

On Intel Macs, reproduced the bug in ANGLE's Metal backend when the
browser was invoked with Metal debug layers and load/store action
validation enabled:
  MTL_DEBUG_LAYER=1 \
  MTL_DEBUG_LAYER_VALIDATE_LOAD_ACTIONS=1 \
  MTL_DEBUG_LAYER_VALIDATE_STORE_ACTIONS=1 \
  MTL_DEBUG_LAYER_VALIDATE_UNRETAINED_RESOURCES=4 [browser...]

A similarly structured test is being added to the ANGLE project.

Minimized test case for http://anglebug.com/6972 .