Allow passing in context creation options in URL
And for showing the `RENDERER`.

This CL allows you to set any context creation parameters
in the URL. Example:

    sometest.html?antialias=false&powerPreference=high-performance

`showRenderer` which will print the renderer, unmasked
if the `WEBGL_debug_renderer_info` exists. Example:

    sometest.html?showRenderer=true

These are useful for testing on dual GPU systems and verifying
that the correct GPU is being tested.

`runUntilFail` which runs the test until it faisl. Useful for testing
intermittent failures.