Initial conformance tests for WEBGL_shader_pixel_local_storage
Creates a set of tests to check the WebGL behavior not found in the
ANGLE_shader_pixel_local_storage specification, as well as providing
some scaffolding to verify the ANGLE extension is properly hooked up.

If these all look good, we can start porting tests from the ANGLE
extension next.