Canvases should reset their contents when width or height is set even if the sizes are the same.
Note: Canvas 2D does this. There are various issues related to
requestAnimationFrame and ResizeObserver that this affects. It
would probably be best of all the APIs did the same thing so that
the workarounds for resizing the canvas are the same across APIs

Safari passes this test, Firefox and Chrome fail