Tolerate throws in wasm/* tests, add 2gb-in-4gb variants.
For browsers that don't support this large of wasm memory sizes,
consider as PASS, so that browser features don't block webgl
conformance.
For UAs that throw on use of large TypedArrays, FAIL but safely.

Also, Firefox *does* support ~4GB wasm memory sizes, but currently
throws when >2GB TypedArray views are passed to entrypoints. However,
Firefox does support <2GB-sized views of >2GB ArrayBuffers, so add tests
for that.
This reflects the ability for apps to use large WASM heaps, so long as
only smaller views are passed to webgl.