Remove MAX_COLOR_ATTACHMENTS_WITH_ACTIVE_PIXEL_LOCAL_STORAGE enum.
Remove this enum from the WEBGL_shader_pixel_local_storage extension,
and renumber the remaining enums, both of which have already been done
in the underlying ANGLE extension.

Remove the use of this enum from the pixel local storage CTS test as
well. The new test passes on Chromium with all of these changes made.

Associated with https://anglebug.com/40096838 .